refactor: extract CLI recipe evidence helpers
